+++ src/testutil.rs
@@ -1,4 +1,3 @@
-use crate::character::Character;
 use num_traits::Zero;
 
 /// Build a text for tests using a generator function `gen`.
@@ -33,8 +32,8 @@ where
     let text = text.as_ref();
     let n = text.len();
     let suffixes = (0..n).map(|i| &text[i..n]).collect::<Vec<_>>();
-    let mut sa = (0..(suffixes.len() as usize)).collect::<Vec<_>>();
-    sa.sort_by_key(|i| &suffixes[*i as usize]);
+    let mut sa = (0..suffixes.len()).collect::<Vec<_>>();
+    sa.sort_by_key(|i| &suffixes[*i]);
     sa
 }
 
@@ -42,7 +41,7 @@ where
 pub fn build_inv_suffix_array(suffix_array: &[usize]) -> Vec<usize> {
     let mut isa = vec![0; suffix_array.len()];
     for (p, &i) in suffix_array.iter().enumerate() {
-        isa[i as usize] = p as usize;
+        isa[i] = p;
     }
     isa
 }
